Fix types for trt 10

+++ csrc/mmdeploy/backend_ops/tensorrt/gather_topk/gather_topk_kernel.cu
@@ -23,8 +23,8 @@ __global__ void gather_topk_kernel(const scalar_t* input, const int* indices, sc
 }
 
 template <typename scalar_t>
-void gather_topk_impl(const scalar_t* input, const int* indices, const int* dims, int nbDims,
-                      const int* indices_dims, int indice_nbDims, scalar_t* output,
+void gather_topk_impl(const scalar_t* input, const int* indices, const int64_t* dims, int nbDims,
+                      const int64_t* indices_dims, int indice_nbDims, scalar_t* output,
                       cudaStream_t stream) {
   int batch = 1;
   for (int i = 0; i < indice_nbDims - 1; ++i) batch *= dims[i];
@@ -37,10 +37,10 @@ void gather_topk_impl(const scalar_t* input, const int* indices, const int* dims
                                                                   num_input, num_indices, channel);
 }
 
-template void gather_topk_impl<float>(const float* input, const int* indices, const int* dims,
-                                      int nbDims, const int* indices_dims, int indice_nbDims,
+template void gather_topk_impl<float>(const float* input, const int* indices, const int64_t* dims,
+                                      int nbDims, const int64_t* indices_dims, int indice_nbDims,
                                       float* output, cudaStream_t stream);
 
-template void gather_topk_impl<int32_t>(const int32_t* input, const int* indices, const int* dims,
-                                        int nbDims, const int* indices_dims, int indice_nbDims,
+template void gather_topk_impl<int32_t>(const int32_t* input, const int* indices, const int64_t* dims,
+                                        int nbDims, const int64_t* indices_dims, int indice_nbDims,
                                         int32_t* output, cudaStream_t stream);
