+++ csrc/mmdeploy/backend_ops/tensorrt/grid_sampler/trt_grid_sampler_kernel.cu
@@ -348,7 +348,7 @@ __global__ void grid_sampler_3d_kernel(const int nthreads, const scalar_t *input
   }
 }
 
-void create_desc(const int *dims, int nb_dims, TensorDesc &desc) {
+void create_desc(const int64_t *dims, int nb_dims, TensorDesc &desc) {
   memcpy(&desc.shape[0], dims, sizeof(int) * nb_dims);
   desc.stride[nb_dims - 1] = 1;
   for (int i = nb_dims - 2; i >= 0; --i) {
@@ -357,8 +357,8 @@ void create_desc(const int *dims, int nb_dims, TensorDesc &desc) {
 }
 
 template <typename T>
-void grid_sample(T *output, const T *input, const T *grid, int *output_dims, int *input_dims,
-                 int *grid_dims, int nb_dims, GridSamplerInterpolation interp,
+void grid_sample(T *output, const T *input, const T *grid, int64_t *output_dims, int64_t *input_dims,
+                 int64_t *grid_dims, int nb_dims, GridSamplerInterpolation interp,
                  GridSamplerPadding padding, bool align_corners, cudaStream_t stream) {
   TensorDesc input_desc;
   create_desc(input_dims, nb_dims, input_desc);
@@ -391,6 +391,6 @@ void grid_sample(T *output, const T *input, const T *grid, int *output_dims, int
 }
 
 template void grid_sample<float>(float *output, const float *input, const float *grid,
-                                 int *output_dims, int *input_dims, int *grid_dims, int nb_dims,
+                                 int64_t *output_dims, int64_t *input_dims, int64_t *grid_dims, int nb_dims,
                                  GridSamplerInterpolation interp, GridSamplerPadding padding,
                                  bool align_corners, cudaStream_t stream);
